+++ src/components/button/button.ts
@@ -2,7 +2,7 @@ import { customElement, property, query, state } from 'lit/decorators.js';
 import { classMap } from 'lit/directives/class-map.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { html, literal } from 'lit/static-html.js';
-import { FormSubmitController } from '../../internal/form';
+import { FormControlController } from '../../internal/form';
 import ShoelaceElement from '../../internal/shoelace-element';
 import { HasSlotController } from '../../internal/slot';
 import { watch } from '../../internal/watch';
@@ -39,7 +39,7 @@ import type { CSSResultGroup } from 'lit';
 export default class SlButton extends ShoelaceElement implements ShoelaceFormControl {
   static styles: CSSResultGroup = styles;
 
-  private readonly formSubmitController = new FormSubmitController(this, {
+  private readonly formControlController = new FormControlController(this, {
     form: input => {
       // Buttons support a form attribute that points to an arbitrary form, so if this attribute it set we need to query
       // the form from the same root using its id
@@ -141,7 +141,7 @@ export default class SlButton extends ShoelaceElement implements ShoelaceFormCon
 
   firstUpdated() {
     if (this.isButton()) {
-      this.invalid = !(this.button as HTMLButtonElement).checkValidity();
+      this.formControlController.updateValidity();
     }
   }
 
@@ -163,11 +163,11 @@ export default class SlButton extends ShoelaceElement implements ShoelaceFormCon
     }
 
     if (this.type === 'submit') {
-      this.formSubmitController.submit(this);
+      this.formControlController.submit(this);
     }
 
     if (this.type === 'reset') {
-      this.formSubmitController.reset(this);
+      this.formControlController.reset(this);
     }
   }
 
@@ -181,10 +181,9 @@ export default class SlButton extends ShoelaceElement implements ShoelaceFormCon
 
   @watch('disabled', { waitUntilFirstUpdate: true })
   handleDisabledChange() {
-    // Disabled form controls are always valid, so we need to recheck validity when the state changes
     if (this.isButton()) {
-      this.button.disabled = this.disabled;
-      this.invalid = !(this.button as HTMLButtonElement).checkValidity();
+      // Disabled form controls are always valid
+      this.formControlController.setValidity(this.disabled);
     }
   }
 
@@ -225,7 +224,7 @@ export default class SlButton extends ShoelaceElement implements ShoelaceFormCon
   setCustomValidity(message: string) {
     if (this.isButton()) {
       (this.button as HTMLButtonElement).setCustomValidity(message);
-      this.invalid = !(this.button as HTMLButtonElement).checkValidity();
+      this.formControlController.updateValidity();
     }
   }