+++ src/internal/form.ts
@@ -21,7 +21,7 @@ const userInteractedControls: WeakMap<ShoelaceFormControl, boolean> = new WeakMa
 //
 const reportValidityOverloads: WeakMap<HTMLFormElement, () => boolean> = new WeakMap();
 
-export interface FormSubmitControllerOptions {
+export interface FormControlControllerOptions {
   /** A function that returns the form containing the form control. */
   form: (input: ShoelaceFormControl) => HTMLFormElement | null;
   /** A function that returns the form control's name, which will be submitted with the form data. */
@@ -41,12 +41,12 @@ export interface FormSubmitControllerOptions {
   setValue: (input: ShoelaceFormControl, value: unknown) => void;
 }
 
-export class FormSubmitController implements ReactiveController {
+export class FormControlController implements ReactiveController {
   host: ShoelaceFormControl & ReactiveControllerHost;
   form?: HTMLFormElement | null;
-  options: FormSubmitControllerOptions;
+  options: FormControlControllerOptions;
 
-  constructor(host: ReactiveControllerHost & ShoelaceFormControl, options?: Partial<FormSubmitControllerOptions>) {
+  constructor(host: ReactiveControllerHost & ShoelaceFormControl, options?: Partial<FormControlControllerOptions>) {
     (this.host = host).addController(this);
     this.options = {
       form: input => input.closest('form'),
@@ -112,37 +112,12 @@ export class FormSubmitController implements ReactiveController {
   }
 
   hostUpdated() {
-    //
-    // We're mapping the following "states" to data attributes. In the future, we can use ElementInternals.states to
-    // create a similar mapping, but instead of [data-invalid] it will look like :--invalid.
-    //
-    // See this RFC for more details: https://github.com/shoelace-style/shoelace/issues/1011
-    //
-    const host = this.host;
-    const hasInteracted = Boolean(userInteractedControls.get(host));
-    const invalid = Boolean(host.invalid);
-    const required = Boolean(host.required);
-
-    if (this.form?.noValidate) {
-      // Form validation is disabled, remove the attributes
-      host.removeAttribute('data-required');
-      host.removeAttribute('data-optional');
-      host.removeAttribute('data-invalid');
-      host.removeAttribute('data-valid');
-      host.removeAttribute('data-user-invalid');
-      host.removeAttribute('data-user-valid');
-    } else {
-      // Form validation is enabled, set the attributes
-      host.toggleAttribute('data-required', required);
-      host.toggleAttribute('data-optional', !required);
-      host.toggleAttribute('data-invalid', invalid);
-      host.toggleAttribute('data-valid', !invalid);
-      host.toggleAttribute('data-user-invalid', invalid && hasInteracted);
-      host.toggleAttribute('data-user-valid', !invalid && hasInteracted);
+    if (this.host.hasUpdated) {
+      this.setValidity(this.host.checkValidity());
     }
   }
 
-  handleFormData(event: FormDataEvent) {
+  private handleFormData(event: FormDataEvent) {
     const disabled = this.options.disabled(this.host);
     const name = this.options.name(this.host);
     const value = this.options.value(this.host);
@@ -162,7 +137,7 @@ export class FormSubmitController implements ReactiveController {
     }
   }
 
-  handleFormSubmit(event: Event) {
+  private handleFormSubmit(event: Event) {
     const disabled = this.options.disabled(this.host);
     const reportValidity = this.options.reportValidity;
 
@@ -179,17 +154,17 @@ export class FormSubmitController implements ReactiveController {
     }
   }
 
-  handleFormReset() {
+  private handleFormReset() {
     this.options.setValue(this.host, this.options.defaultValue(this.host));
     this.setUserInteracted(this.host, false);
   }
 
-  async handleUserInput() {
+  private async handleUserInput() {
     await this.host.updateComplete;
     this.setUserInteracted(this.host, true);
   }
 
-  reportFormValidity() {
+  private reportFormValidity() {
     //
     // Shoelace form controls work hard to act like regular form controls. They support the Constraint Validation API
     // and its associated methods such as setCustomValidity() and reportValidity(). However, the HTMLFormElement also
@@ -219,12 +194,12 @@ export class FormSubmitController implements ReactiveController {
     return true;
   }
 
-  setUserInteracted(el: ShoelaceFormControl, hasInteracted: boolean) {
+  private setUserInteracted(el: ShoelaceFormControl, hasInteracted: boolean) {
     userInteractedControls.set(el, hasInteracted);
     el.requestUpdate();
   }
 
-  doAction(type: 'submit' | 'reset', invoker?: HTMLInputElement | SlButton) {
+  private doAction(type: 'submit' | 'reset', invoker?: HTMLInputElement | SlButton) {
     if (this.form) {
       const button = document.createElement('button');
       button.type = type;
@@ -263,5 +238,48 @@ export class FormSubmitController implements ReactiveController {
     // Calling form.submit() bypasses the submit event and constraint validation. To prevent this, we can inject a
     // native submit button into the form, "click" it, then remove it to simulate a standard form submission.
     this.doAction('submit', invoker);
+  }
+
+  /**
+   * Synchronously sets the form control's validity. Call this when you know the future validity but need to update
+   * the host element immediately, i.e. before Lit updates the component in the next update.
+   */
+  setValidity(isValid: boolean) {
+    const host = this.host;
+    const hasInteracted = Boolean(userInteractedControls.get(host));
+    const required = Boolean(host.required);
+
+    //
+    // We're mapping the following "states" to data attributes. In the future, we can use ElementInternals.states to
+    // create a similar mapping, but instead of [data-invalid] it will look like :--invalid.
+    //
+    // See this RFC for more details: https://github.com/shoelace-style/shoelace/issues/1011
+    //
+    if (this.form?.noValidate) {
+      // Form validation is disabled, remove the attributes
+      host.removeAttribute('data-required');
+      host.removeAttribute('data-optional');
+      host.removeAttribute('data-invalid');
+      host.removeAttribute('data-valid');
+      host.removeAttribute('data-user-invalid');
+      host.removeAttribute('data-user-valid');
+    } else {
+      // Form validation is enabled, set the attributes
+      host.toggleAttribute('data-required', required);
+      host.toggleAttribute('data-optional', !required);
+      host.toggleAttribute('data-invalid', !isValid);
+      host.toggleAttribute('data-valid', isValid);
+      host.toggleAttribute('data-user-invalid', !isValid && hasInteracted);
+      host.toggleAttribute('data-user-valid', isValid && hasInteracted);
+    }
+  }
+
+  /**
+   * Updates the form control's validity based on the current value of `host.checkValidity()`. Call this when anything
+   * that affects constraint validation changes so the component receives the correct validity states.
+   */
+  updateValidity() {
+    const host = this.host;
+    this.setValidity(host.checkValidity());
   }
 }